advanced_tensorflow/pyproject.toml
[tool.poetry]
name = "advanced_tensorflow"
version = "0.1.0"
description = ""
authors = ["Jonathas <jhsalves@gmail.com>"]

[tool.poetry.dependencies]
python = ">=3.8,<3.10"
tensorflow-cpu = "^2.7.0"
pandas = "^1.3.4"
flwr = "^0.17.0"
seaborn = "^0.11.2"
sklearn = "^0.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
